+++ reddit_api_kernel/endpoint.mli
@@ -314,11 +314,6 @@ type 'a t =
 
 val map : 'a t -> f:('a -> 'b) -> 'b t
 
-type 'a with_param_override :=
-  ?param_list_override:((string * string list) list -> (string * string list) list)
-  -> unit
-  -> 'a
-
 (** A value of type [_ with_listing_params] is a function with optional
     arguments representing Reddit's "listing" pagination protocol.
 
@@ -331,13 +326,13 @@ type 'a with_listing_params :=
 
 (** {2 Account } *)
 
-val me : User.t t with_param_override
-val karma : Karma_list.t t with_param_override
-val trophies : Award.t list t with_param_override
-val friends : User_list.t t with_param_override with_listing_params
-val blocked : User_list.t t with_param_override with_listing_params
-val messaging : User_list.t t with_param_override with_listing_params
-val trusted : User_list.t t with_param_override with_listing_params
+val me : User.t t
+val karma : Karma_list.t t
+val trophies : Award.t list t
+val friends : (unit -> User_list.t t) with_listing_params
+val blocked : (unit -> User_list.t t) with_listing_params
+val messaging : (unit -> User_list.t t) with_listing_params
+val trusted : (unit -> User_list.t t) with_listing_params
 
 (** {2 Flair } *)
 
@@ -347,51 +342,45 @@ val select_flair
   -> ?flair_template_id:Uuid.t
   -> ?text:string
   -> ?text_color:Color.t
+  -> unit
   -> subreddit:Subreddit_name.t
   -> target:Flair_target.t
-  -> unit t with_param_override
+  -> unit t
 
 (** {2 Links and comments } *)
 
 val add_comment
   :  ?return_rtjson:bool
   -> ?richtext_json:Json.t
+  -> unit
   -> parent:[< `Link of Link.Id.t | `Comment of Comment.Id.t | `Message of Message.Id.t ]
   -> text:string
-  -> Thing.Comment.t t with_param_override
+  -> Thing.Comment.t t
 
-val delete
-  :  id:[< `Link of Link.Id.t | `Comment of Comment.Id.t ]
-  -> unit t with_param_override
+val delete : id:[< `Link of Link.Id.t | `Comment of Comment.Id.t ] -> unit t
 
 val edit
   :  ?return_rtjson:bool
   -> ?richtext_json:Json.t
+  -> unit
   -> id:[< `Link of Link.Id.t | `Comment of Comment.Id.t ]
   -> text:string
-  -> [> `Link of Link.t | `Comment of Comment.t ] t with_param_override
+  -> [> `Link of Link.t | `Comment of Comment.t ] t
 
-val hide : links:Link.Id.t list -> unit t with_param_override
-val unhide : links:Link.Id.t list -> unit t with_param_override
-
-val lock
-  :  id:[< `Link of Link.Id.t | `Comment of Comment.Id.t ]
-  -> unit t with_param_override
-
-val unlock
-  :  id:[< `Link of Link.Id.t | `Comment of Comment.Id.t ]
-  -> unit t with_param_override
-
-val mark_nsfw : link:Link.Id.t -> unit t with_param_override
-val unmark_nsfw : link:Link.Id.t -> unit t with_param_override
+val hide : links:Link.Id.t list -> unit t
+val unhide : links:Link.Id.t list -> unit t
+val lock : id:[< `Link of Link.Id.t | `Comment of Comment.Id.t ] -> unit t
+val unlock : id:[< `Link of Link.Id.t | `Comment of Comment.Id.t ] -> unit t
+val mark_nsfw : link:Link.Id.t -> unit t
+val unmark_nsfw : link:Link.Id.t -> unit t
 
 val more_children
   :  ?limit_children:bool
+  -> unit
   -> link:Link.Id.t
   -> more_comments:More_comments.Details.By_children.t
   -> sort:Comment_sort.t
   -> [ `Comment of Comment.t | `More_comments of More_comments.t ] list t
-     with_param_override
 
 val report
   :  ?from_modmail:bool
@@ -402,6 +391,7 @@ val report
   -> ?rule_reason:string
   -> ?site_reason:string
   -> ?sr_name:string
+  -> unit
   -> target:
        [< `Link of Link.Id.t
        | `Comment of Comment.Id.t
@@ -409,47 +399,37 @@ val report
        | `Modmail_conversation of Modmail_conversation.Id.t
        ]
   -> reason:string
-  -> unit t with_param_override
+  -> unit t
 
-val report_award
-  :  award_id:string
-  -> (Cohttp.Response.t * Cohttp.Body.t) t with_param_override
+val report_award : award_id:string -> (Cohttp.Response.t * Cohttp.Body.t) t
 
 val save
   :  ?category:string
+  -> unit
   -> id:[< `Link of Link.Id.t | `Comment of Comment.Id.t ]
-  -> unit t with_param_override
+  -> unit t
 
-val unsave
-  :  id:[< `Link of Link.Id.t | `Comment of Comment.Id.t ]
-  -> unit t with_param_override
-
-val saved_categories : (Cohttp.Response.t * Cohttp.Body.t) t with_param_override
+val unsave : id:[< `Link of Link.Id.t | `Comment of Comment.Id.t ] -> unit t
+val saved_categories : (Cohttp.Response.t * Cohttp.Body.t) t
 
 val send_replies
   :  id:[< `Link of Link.Id.t | `Comment of Comment.Id.t ]
   -> enabled:bool
-  -> unit t with_param_override
+  -> unit t
 
-val set_contest_mode : link:Link.Id.t -> enabled:bool -> unit t with_param_override
+val set_contest_mode : link:Link.Id.t -> enabled:bool -> unit t
 
 val set_subreddit_sticky
   :  ?to_profile:bool
+  -> unit
   -> link:Link.Id.t
   -> sticky_state:Sticky_state.t
-  -> unit t with_param_override
+  -> unit t
 
-val set_suggested_sort
-  :  link:Link.Id.t
-  -> sort:Comment_sort.t option
-  -> unit t with_param_override
-
-val spoiler : link:Link.Id.t -> unit t with_param_override
-val unspoiler : link:Link.Id.t -> unit t with_param_override
-
-val store_visits
-  :  links:Link.Id.t list
-  -> (Cohttp.Response.t * Cohttp.Body.t) t with_param_override
+val set_suggested_sort : link:Link.Id.t -> sort:Comment_sort.t option -> unit t
+val spoiler : link:Link.Id.t -> unit t
+val unspoiler : link:Link.Id.t -> unit t
+val store_visits : links:Link.Id.t list -> (Cohttp.Response.t * Cohttp.Body.t) t
 
 val submit
   :  ?ad:bool
@@ -463,16 +443,18 @@ val submit
   -> ?event_start:Time.t
   -> ?event_end:Time.t
   -> ?event_tz:string
+  -> unit
   -> subreddit:Subreddit_name.t
   -> title:string
   -> kind:Link_kind.t
-  -> (Link.Id.t * Uri.t) t with_param_override
+  -> (Link.Id.t * Uri.t) t
 
 val vote
   :  ?rank:int
+  -> unit
   -> direction:Vote_direction.t
   -> target:[< `Link of Link.Id.t | `Comment of Comment.Id.t ]
-  -> unit t with_param_override
+  -> unit t
 
 val info
   :  Info_query.t
@@ -482,15 +464,11 @@ val info
      ]
      list
      t
-     with_param_override
 
 (** {2 Listings } *)
 
-val best
-  : (?include_categories:bool -> Link.t Listing.t t with_param_override)
-    with_listing_params
-
-val links_by_id : links:Link.Id.t list -> Link.t Listing.t t with_param_override
+val best : (?include_categories:bool -> unit -> Link.t Listing.t t) with_listing_params
+val links_by_id : links:Link.Id.t list -> Link.t Listing.t t
 
 val comments
   :  ?subreddit:Subreddit_name.t
@@ -503,103 +481,107 @@ val comments
   -> ?sort:Comment_sort.t
   -> ?threaded:bool
   -> ?truncate:int
+  -> unit
   -> link:Link.Id.t
-  -> Comment_response.t t with_param_override
+  -> Comment_response.t t
 
 val duplicates
   : (?crossposts_only:bool
      -> ?sort:Duplicate_sort.t
+     -> unit
      -> link:Link.Id.t
-     -> Link.t Listing.t t with_param_override)
+     -> Link.t Listing.t t)
     with_listing_params
 
 val hot
   : (?location:string
      -> ?include_categories:bool
      -> ?subreddit:Subreddit_name.t
-     -> Link.t Listing.t t with_param_override)
+     -> unit
+     -> Link.t Listing.t t)
     with_listing_params
 
 val new_
   : (?include_categories:bool
      -> ?subreddit:Subreddit_name.t
-     -> Link.t Listing.t t with_param_override)
+     -> unit
+     -> Link.t Listing.t t)
     with_listing_params
 
 val rising
   : (?include_categories:bool
      -> ?subreddit:Subreddit_name.t
-     -> Link.t Listing.t t with_param_override)
+     -> unit
+     -> Link.t Listing.t t)
     with_listing_params
 
 val top
   : (?since:Historical_span.t
      -> ?include_categories:bool
      -> ?subreddit:Subreddit_name.t
-     -> Link.t Listing.t t with_param_override)
+     -> unit
+     -> Link.t Listing.t t)
     with_listing_params
 
 val controversial
   : (?since:Historical_span.t
      -> ?include_categories:bool
      -> ?subreddit:Subreddit_name.t
-     -> Link.t Listing.t t with_param_override)
+     -> unit
+     -> Link.t Listing.t t)
     with_listing_params
 
-val random : ?subreddit:Subreddit_name.t -> Link.Id.t t with_param_override
+val random : ?subreddit:Subreddit_name.t -> unit -> Link.Id.t t
 
 (** {2 Private messages } *)
 
-val block_author
-  :  id:[< `Comment of Comment.Id.t | `Message of Message.Id.t ]
-  -> unit t with_param_override
-
-val collapse_message : messages:Message.Id.t list -> unit t with_param_override
-val uncollapse_message : messages:Message.Id.t list -> unit t with_param_override
+val block_author : id:[< `Comment of Comment.Id.t | `Message of Message.Id.t ] -> unit t
+val collapse_message : messages:Message.Id.t list -> unit t
+val uncollapse_message : messages:Message.Id.t list -> unit t
 
 val compose_message
   :  ?g_recaptcha_response:string
   -> ?from_subreddit:Subreddit_name.t
+  -> unit
   -> to_:Username.t
   -> subject:string
   -> text:string
-  -> unit t with_param_override
+  -> unit t
 
-val delete_message
-  :  message:Message.Id.t
-  -> (Cohttp.Response.t * Cohttp.Body.t) t with_param_override
-
-val read_message : messages:Message.Id.t list -> unit t with_param_override
-val unread_message : messages:Message.Id.t list -> unit t with_param_override
+val delete_message : message:Message.Id.t -> (Cohttp.Response.t * Cohttp.Body.t) t
+val read_message : messages:Message.Id.t list -> unit t
+val unread_message : messages:Message.Id.t list -> unit t
 
 val inbox
   : (?include_categories:bool
      -> ?mid:string
+     -> unit
      -> mark_read:bool
-     -> Inbox_item.t Listing.t t with_param_override)
+     -> Inbox_item.t Listing.t t)
     with_listing_params
 
 val unread
   : (?include_categories:bool
      -> ?mid:string
+     -> unit
      -> mark_read:bool
-     -> Inbox_item.t Listing.t t with_param_override)
+     -> Inbox_item.t Listing.t t)
     with_listing_params
 
 val sent
-  : (?include_categories:bool -> ?mid:string -> Message.t Listing.t t with_param_override)
+  : (?include_categories:bool -> ?mid:string -> unit -> Message.t Listing.t t)
     with_listing_params
 
 val comment_replies
   : (?include_categories:bool
      -> ?mid:string
+     -> unit
      -> mark_read:bool
-     -> Inbox_item.Comment.t Listing.t t with_param_override)
+     -> Inbox_item.Comment.t Listing.t t)
     with_listing_params
 
 val subreddit_comments
-  : (subreddit:Subreddit_name.t -> Comment.t Listing.t t with_param_override)
-    with_listing_params
+  : (subreddit:Subreddit_name.t -> Comment.t Listing.t t) with_listing_params
 
 (** {2 Moderation } *)
 
@@ -607,76 +589,71 @@ val log
   : (?mod_filter:Mod_filter.t
      -> ?subreddit:Subreddit_name.t
      -> ?type_:string
-     -> Mod_action.t Listing.t t with_param_override)
+     -> unit
+     -> Mod_action.t Listing.t t)
     with_listing_params
 
 val reports
   : (?location:string
      -> ?only:Links_or_comments.t
      -> ?subreddit:Subreddit_name.t
-     -> [ `Link of Link.t | `Comment of Comment.t ] Listing.t t with_param_override)
+     -> unit
+     -> [ `Link of Link.t | `Comment of Comment.t ] Listing.t t)
     with_listing_params
 
 val spam
   : (?location:string
      -> ?only:Links_or_comments.t
      -> ?subreddit:Subreddit_name.t
-     -> [ `Link of Link.t | `Comment of Comment.t ] Listing.t t with_param_override)
+     -> unit
+     -> [ `Link of Link.t | `Comment of Comment.t ] Listing.t t)
     with_listing_params
 
 val modqueue
   : (?location:string
      -> ?only:Links_or_comments.t
      -> ?subreddit:Subreddit_name.t
-     -> [ `Link of Link.t | `Comment of Comment.t ] Listing.t t with_param_override)
+     -> unit
+     -> [ `Link of Link.t | `Comment of Comment.t ] Listing.t t)
     with_listing_params
 
 val unmoderated
   : (?location:string
      -> ?only:Links_or_comments.t
      -> ?subreddit:Subreddit_name.t
-     -> [ `Link of Link.t | `Comment of Comment.t ] Listing.t t with_param_override)
+     -> unit
+     -> [ `Link of Link.t | `Comment of Comment.t ] Listing.t t)
     with_listing_params
 
 val edited
   : (?location:string
      -> ?only:Links_or_comments.t
      -> ?subreddit:Subreddit_name.t
-     -> [ `Link of Link.t | `Comment of Comment.t ] Listing.t t with_param_override)
+     -> unit
+     -> [ `Link of Link.t | `Comment of Comment.t ] Listing.t t)
     with_listing_params
 
 val accept_moderator_invite
   :  subreddit:Subreddit_name.t
-  -> (Cohttp.Response.t * Cohttp.Body.t) t with_param_override
+  -> (Cohttp.Response.t * Cohttp.Body.t) t
 
-val approve
-  :  id:[< `Link of Link.Id.t | `Comment of Comment.Id.t ]
-  -> unit t with_param_override
-
-val remove
-  :  id:[< `Link of Link.Id.t | `Comment of Comment.Id.t ]
-  -> spam:bool
-  -> unit t with_param_override
+val approve : id:[< `Link of Link.Id.t | `Comment of Comment.Id.t ] -> unit t
+val remove : id:[< `Link of Link.Id.t | `Comment of Comment.Id.t ] -> spam:bool -> unit t
 
 val distinguish
   :  ?sticky:bool
+  -> unit
   -> id:[< `Link of Link.Id.t | `Comment of Comment.Id.t ]
   -> how:How_to_distinguish.t
-  -> [> `Link of Link.t | `Comment of Comment.t ] t with_param_override
+  -> [> `Link of Link.t | `Comment of Comment.t ] t
 
-val ignore_reports
-  :  id:[< `Link of Link.Id.t | `Comment of Comment.Id.t ]
-  -> unit t with_param_override
-
-val unignore_reports
-  :  id:[< `Link of Link.Id.t | `Comment of Comment.Id.t ]
-  -> unit t with_param_override
-
-val leavecontributor : subreddit:Subreddit.Id.t -> unit t with_param_override
-val leavemoderator : subreddit:Subreddit.Id.t -> unit t with_param_override
-val mute_message_author : message:Message.Id.t -> unit t with_param_override
-val unmute_message_author : message:Message.Id.t -> unit t with_param_override
-val stylesheet : subreddit:Subreddit_name.t -> Stylesheet.t t with_param_override
+val ignore_reports : id:[< `Link of Link.Id.t | `Comment of Comment.Id.t ] -> unit t
+val unignore_reports : id:[< `Link of Link.Id.t | `Comment of Comment.Id.t ] -> unit t
+val leavecontributor : subreddit:Subreddit.Id.t -> unit t
+val leavemoderator : subreddit:Subreddit.Id.t -> unit t
+val mute_message_author : message:Message.Id.t -> unit t
+val unmute_message_author : message:Message.Id.t -> unit t
+val stylesheet : subreddit:Subreddit_name.t -> Stylesheet.t t
 
 (** {2 New modmail } *)
 
@@ -686,7 +663,7 @@ val create_modmail_conversation
   -> subreddit:Subreddit_name.t
   -> to_:Modmail_recipient.t
   -> hide_author:bool
-  -> Modmail.Conversation.t t with_param_override
+  -> Modmail.Conversation.t t
 
 (** {2 Search } *)
 
@@ -697,11 +674,11 @@ val search
      -> ?since:Historical_span.t
      -> ?sort:Search_sort.t
      -> ?types:Search_type.Set.t
+     -> unit
      -> query:string
      -> (Thing.Link.t Listing.t option
         * [ `Subreddit of Thing.Subreddit.t | `User of Thing.User.t ] Listing.t option)
-        t
-        with_param_override)
+        t)
     with_listing_params
 
 (** {2 Subreddits } *)
@@ -709,61 +686,69 @@ val search
 val banned
   : (?include_categories:bool
      -> ?user:Username.t
+     -> unit
      -> subreddit:Subreddit_name.t
-     -> Relationship.Ban.t Listing.t t with_param_override)
+     -> Relationship.Ban.t Listing.t t)
     with_listing_params
 
 val muted
   : (?include_categories:bool
      -> ?user:Username.t
+     -> unit
      -> subreddit:Subreddit_name.t
-     -> Relationship.Mute.t Listing.t t with_param_override)
+     -> Relationship.Mute.t Listing.t t)
     with_listing_params
 
 val wiki_banned
   : (?include_categories:bool
      -> ?user:Username.t
+     -> unit
      -> subreddit:Subreddit_name.t
-     -> Relationship.Ban.t Listing.t t with_param_override)
+     -> Relationship.Ban.t Listing.t t)
     with_listing_params
 
 val contributors
   : (?include_categories:bool
      -> ?user:Username.t
+     -> unit
      -> subreddit:Subreddit_name.t
-     -> Relationship.Contributor.t Listing.t t with_param_override)
+     -> Relationship.Contributor.t Listing.t t)
     with_listing_params
 
 val wiki_contributors
   : (?include_categories:bool
      -> ?user:Username.t
+     -> unit
      -> subreddit:Subreddit_name.t
-     -> Relationship.Contributor.t Listing.t t with_param_override)
+     -> Relationship.Contributor.t Listing.t t)
     with_listing_params
 
 val moderators
   : (?include_categories:bool
      -> ?user:Username.t
+     -> unit
      -> subreddit:Subreddit_name.t
-     -> Relationship.Moderator.t Listing.t t with_param_override)
+     -> Relationship.Moderator.t Listing.t t)
     with_listing_params
 
 val delete_subreddit_image
   :  subreddit:Subreddit_name.t
   -> image:Subreddit_image.t
-  -> unit t with_param_override
+  -> unit t
 
 val search_subreddits_by_name
   :  ?exact:bool
   -> ?include_over_18:bool
   -> ?include_unadvertisable:bool
+  -> unit
   -> query:string
-  -> Subreddit_name.t list t with_param_override
+  -> Subreddit_name.t list t
 
 val create_or_edit_subreddit
   :  ?comment_score_hide_mins:int
   -> ?wiki_edit_age:int
   -> ?wiki_edit_karma:int
+  -> unit
   -> all_original_content:bool
   -> allow_discovery:bool
   -> allow_images:bool
@@ -803,155 +788,146 @@ val create_or_edit_subreddit
   -> title:string
   -> type_:Subreddit_type.t
   -> wiki_mode:Wiki_mode.t
-  -> (Cohttp.Response.t * Cohttp.Body.t) t with_param_override
+  -> (Cohttp.Response.t * Cohttp.Body.t) t
 
-val submit_text : subreddit:Subreddit_name.t -> Submit_text.t t with_param_override
+val submit_text : subreddit:Subreddit_name.t -> Submit_text.t t
 
 val subreddit_autocomplete
   :  ?limit:int
   -> ?include_categories:bool
   -> ?include_over_18:bool
   -> ?include_profiles:bool
+  -> unit
   -> query:string
-  -> Subreddit.t Listing.t t with_param_override
+  -> Subreddit.t Listing.t t
 
 val set_subreddit_stylesheet
   :  ?reason:string
+  -> unit
   -> subreddit:Subreddit_name.t
   -> stylesheet_contents:string
-  -> unit t with_param_override
+  -> unit t
 
 val subscribe
   :  ?skip_initial_defaults:bool
+  -> unit
   -> action:Subscription_action.t
   -> subreddits:Subscription_list.t
-  -> unit t with_param_override
+  -> unit t
 
 val search_users
-  : (?sort:Relevance_or_activity.t
-     -> query:string
-     -> User.t Listing.t t with_param_override)
+  : (?sort:Relevance_or_activity.t -> unit -> query:string -> User.t Listing.t t)
     with_listing_params
 
-val about_subreddit : subreddit:Subreddit_name.t -> Subreddit.t t with_param_override
+val about_subreddit : subreddit:Subreddit_name.t -> Subreddit.t t
 
 val subreddit_settings
   :  ?created:bool
   -> ?location:string
+  -> unit
   -> subreddit:Subreddit_name.t
-  -> Subreddit_settings.t t with_param_override
+  -> Subreddit_settings.t t
 
-val subreddit_rules
-  :  subreddit:Subreddit_name.t
-  -> Subreddit_rules.t t with_param_override
-
-val subreddit_traffic
-  :  subreddit:Subreddit_name.t
-  -> Subreddit_traffic.t t with_param_override
-
-val get_sticky
-  :  ?number:int
-  -> subreddit:Subreddit_name.t
-  -> Link.Id.t t with_param_override
+val subreddit_rules : subreddit:Subreddit_name.t -> Subreddit_rules.t t
+val subreddit_traffic : subreddit:Subreddit_name.t -> Subreddit_traffic.t t
+val get_sticky : ?number:int -> unit -> subreddit:Subreddit_name.t -> Link.Id.t t
 
 val get_subreddits
   : (?include_categories:bool
+     -> unit
      -> relationship:Subreddit_relationship.t
-     -> Subreddit.t Listing.t t with_param_override)
+     -> Subreddit.t Listing.t t)
     with_listing_params
 
 val search_subreddits_by_title_and_description
   : (?show_users:bool
      -> ?sort:Relevance_or_activity.t
+     -> unit
      -> query:string
-     -> Subreddit.t Listing.t t with_param_override)
+     -> Subreddit.t Listing.t t)
     with_listing_params
 
 val list_subreddits
   : (?include_categories:bool
      -> ?show_users:bool
+     -> unit
      -> sort:Subreddit_listing_sort.t
-     -> Subreddit.t Listing.t t with_param_override)
+     -> Subreddit.t Listing.t t)
     with_listing_params
 
 (** {2 Users } *)
 
-val about_user : username:Username.t -> User.t t with_param_override
-val user_trophies : username:Username.t -> Award.t list t with_param_override
+val about_user : username:Username.t -> User.t t
+val user_trophies : username:Username.t -> Award.t list t
 
 val list_user_subreddits
   : (?include_categories:bool
+     -> unit
      -> sort:User_subreddit_sort.t
-     -> Subreddit.t Listing.t t with_param_override)
+     -> Subreddit.t Listing.t t)
     with_listing_params
 
 val add_relationship
-  :  relationship:Relationship_spec.t
-  -> username:Username.t
-  -> duration:Relationship_spec.Duration.t
-  -> ?subreddit:Subreddit_name.t
+  :  ?subreddit:Subreddit_name.t
   -> ?note:string
   -> ?ban_reason:string
   -> ?ban_message:string
   -> ?ban_context:string
-  -> unit t with_param_override
+  -> unit
+  -> relationship:Relationship_spec.t
+  -> username:Username.t
+  -> duration:Relationship_spec.Duration.t
+  -> unit t
 
 val remove_relationship
-  :  relationship:Relationship_spec.t
+  :  ?subreddit:Subreddit_name.t
+  -> unit
+  -> relationship:Relationship_spec.t
   -> username:Username.t
-  -> ?subreddit:Subreddit_name.t
-  -> unit t with_param_override
+  -> unit t
 
 (** {2 Wiki } *)
 
-val add_wiki_editor : page:Wiki_page.Id.t -> user:Username.t -> unit t with_param_override
-
-val remove_wiki_editor
-  :  page:Wiki_page.Id.t
-  -> user:Username.t
-  -> unit t with_param_override
+val add_wiki_editor : page:Wiki_page.Id.t -> user:Username.t -> unit t
+val remove_wiki_editor : page:Wiki_page.Id.t -> user:Username.t -> unit t
 
 val edit_wiki_page
   :  ?previous:Wiki_page.Revision.Id.t
   -> ?reason:string
+  -> unit
   -> content:string
   -> page:Wiki_page.Id.t
-  -> (unit, Wiki_page.Edit_conflict.t) Result.t t with_param_override
+  -> (unit, Wiki_page.Edit_conflict.t) Result.t t
 
 val toggle_wiki_revision_visibility
   :  page:Wiki_page.Id.t
   -> revision:Wiki_page.Revision.Id.t
-  -> [ `Became_hidden | `Became_visible ] t with_param_override
+  -> [ `Became_hidden | `Became_visible ] t
 
-val revert_wiki_page
-  :  page:Wiki_page.Id.t
-  -> revision:Wiki_page.Revision.Id.t
-  -> unit t with_param_override
+val revert_wiki_page : page:Wiki_page.Id.t -> revision:Wiki_page.Revision.Id.t -> unit t
 
 val wiki_discussions
-  : (page:Wiki_page.Id.t -> Link.t Listing.t t with_param_override) with_listing_params
+  : (unit -> page:Wiki_page.Id.t -> Link.t Listing.t t) with_listing_params
 
-val wiki_pages : ?subreddit:Subreddit_name.t -> string list t with_param_override
+val wiki_pages : ?subreddit:Subreddit_name.t -> unit -> string list t
 
 val subreddit_wiki_revisions
-  : (?subreddit:Subreddit_name.t -> Wiki_page.Revision.t Listing.t t with_param_override)
+  : (?subreddit:Subreddit_name.t -> unit -> Wiki_page.Revision.t Listing.t t)
     with_listing_params
 
 val wiki_page_revisions
-  : (page:Wiki_page.Id.t -> Wiki_page.Revision.t Listing.t t with_param_override)
-    with_listing_params
+  : (unit -> page:Wiki_page.Id.t -> Wiki_page.Revision.t Listing.t t) with_listing_params
 
-val wiki_permissions
-  :  page:Wiki_page.Id.t
-  -> Wiki_page.Permissions.t t with_param_override
+val wiki_permissions : page:Wiki_page.Id.t -> Wiki_page.Permissions.t t
 
 val set_wiki_permissions
   :  page:Wiki_page.Id.t
   -> listed:bool
   -> level:Wiki_page.Permissions.Level.t
-  -> Wiki_page.Permissions.t t with_param_override
+  -> Wiki_page.Permissions.t t
 
 val wiki_page
   :  ?compare_revisions:string option * string option
+  -> unit
   -> page:Wiki_page.Id.t
-  -> Wiki_page.t t with_param_override
+  -> Wiki_page.t t